+++ src/anim/LinearAnim.h
@@ -31,7 +31,7 @@ namespace avg {
 class AVG_API LinearAnim: public SimpleAnim {
 public:
     LinearAnim(const boost::python::object& node, const std::string& sAttrName, 
-            double duration,
+            long long duration,
             const boost::python::object& pStartValue, 
             const boost::python::object& pEndValue, 
             bool bUseInt,
